For #171: Reservation duration calculation
- Moved Reservation tests from test_models.py to test_reservation.py
- Added Reservation duration() method and removed duration field
